travis: Run a compatible version of the deployment engine
Before, Travis ran with the release version of the bindings, but the
unreleased bleeding edge version of the deployment engine. This caused
failures whenever the deployment engine was updated in a
non-backwards-compatible way.

This commit updates Travis to pick the deployment engine version based
on the bindings version specified in the package.json.